Add the initial project files
The project is a dressed down Symfony 3.4 project with only CLI
capabilities. Project wise, this project was based on the SP dashboard
project. Inheriting the layered architecture.

A simple ansible provisioning was added, building a bare bones
development environment. This should be improved upon!

The LastLogin entity was also included as this is the cornerstone of
the application.