Rename LastLoginService to InformationService
This confusing service was renamed to a more intuitive servicename. The
previous name was somewhat justified as the information request and
subsequent deprovision request used the last login information to
perform the retrieval actions.

Giving the service the single responsibility of retrieving the user
information also required the rename of the service.