Move InformationResponse factory to application
This logic sort of violates the single responsibility priciple and
also introduces vendor logic into our domain layer. A move of the
factory method to the application layer solves those problems.

The dependency injection needed some minor updates and the test case
that tested the factory mehtod was moved and was renamed.