Simplify the DeprovisionClientCollection
Like the previous commit, the DeprovisionClientCollection was
implementing the DeprovisionClientInterface. With the benefit of
ensuring the implementation of the deprovision and information methods.

The new response collection format however makes it more practical to
let the DeprovisionClientCollection implement its own interface. In
which the only notable change is the change of return type for
deprovision and information. And the removal of the getName method.