Remove user from last login
When the user is removed successfully, his entry in the last login table
can be cleaned up. This was not yet implemented and has been added in
this commit. A new service and command where introduced to help the
deprovision service with the removal of the user from the last login
table.