Add summary on top of the JSON response
To reach this goal the commands get the information response collections
instead of a json representation of the collection. Enabling us to
create a summary.

These information responses are then loaded into the newly created
summary service. This service analyzes the information and returns a
console outputable string with a summary of the action.

The summary service can be set with a context (information/deprovision)
modifying the output that is generated between these modes.