Add the web folder and config
The framework config section was extended with web request/routing
related settings. The security config was also added and will later be
configured with basic auth on the api endpoint.

The web folder was added and filled with the sensible defaults (app.php,
app_dev.php, robots.txt and the config.php.