Removed the ApiControllerTest
The deprovision api itself is not covered by a web test. This was not
easily done in the time we had during development of the api. Test
coverage of the used business logic is up to par.

The difficulty lies in the similarities between the newly introduced web
test and the existing console command tests. These tests must bootstrap
the Symfony environment and override certain services for the test
framework to work correctly.

See: https://www.pivotaltracker.com/story/show/158816352