Add database connection failed error handling
When database connection fails, the connection exception from the ORM is
catched. And a home made exception is thrown on top of it. Resulting in
a user friendly output for the console command.

Before it would read:
`An exception occurred in driver: SQLSTATE[HY000] [2002] Connection refused`

And now as requested in the user story:
`No connection possible to the database`

Point 2 of https://www.pivotaltracker.com/story/show/176168526/comments/229746463