Create a last_login test data seeder
This seeder can create a given amount of last_login entries. A date
offset must be specified to control the date of the users last login
time.

This script is usefull when you want to simulate larger batch deprovision
actions. But note that the actual users probably do not exist on the
platform.